% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibration_cpp}
\alias{calibration_cpp}
\title{Calibrating the balancing factors}
\arguments{
\item{cost_fun}{The matrix representing the cost function f(D), where D is the distance matrix.}

\item{O}{A numeric column vector of weights associated to the origin, the production constraint, usually the outbound flow.}

\item{D}{A numeric row vector of destination weights, the attraction constraint.}

\item{delta}{The error term we can tolerate on the convergence of the values.}
}
\value{
A list object containing the A and B vectors of coefficients
and a vector e to verify convergence speed.
}
\description{
Function to calibrate the A and B coefficients of the gravity model through an iterative process.
To avoid infinite loop, the maximum number of iterations is fixed to 50, it usually takes around 5-7 iterations.
to converge.
}
\examples{
cost_fun = matrix(data = c(1,2,3,4,5,6,7,8,9), nrow = 3)
O = c(1,2,3)
D = c(3,2,1)

}
