% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_found_openmp}
\alias{cpp_found_openmp}
\title{Function to check availability of OPENMP to run in parallel.
If openmp is found, this function returns TRUE}
\usage{
cpp_found_openmp()
}
\description{
Function to check availability of OPENMP to run in parallel.
If openmp is found, this function returns TRUE
}
