% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distance_test}
\alias{distance_test}
\title{distance_test}
\format{
\subsection{\code{distance_test}}{

A 983x983 matrix of distances between MSOAs in London.
Computed using the London road network from OpenStreetMap and the cppRouting package.
}
}
\source{
Ivann Schlosser, 2022
}
\usage{
distance_test
}
\description{
distance_test
}
\keyword{datasets}
