% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flows_london}
\alias{flows_london}
\title{flows_london}
\format{
\subsection{\code{flows_london}}{

A data.table with flows information
}
}
\source{
ONS, Office for National Statistics, 2011
}
\usage{
flows_london
}
\description{
flows_london
}
\keyword{datasets}
