% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flows_test}
\alias{flows_test}
\title{flows_test}
\format{
\subsection{\code{flows_test}}{

A matrix of size 983x983 containing flows of users using walking or cycling as the main method of commute.
}
}
\source{
UK Census, 2011
}
\usage{
flows_test
}
\description{
flows_test
}
\keyword{datasets}
