% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london_msoa}
\alias{london_msoa}
\title{london_msoa}
\format{
\subsection{\code{london_msoa}}{

A data.table with London MSOA, their centroids and geometries
}
}
\source{
ONS, Office for National Statistics, 2011
}
\usage{
london_msoa
}
\description{
london_msoa
}
\keyword{datasets}
