% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_model_cpp}
\alias{run_model_cpp}
\title{Run model}
\arguments{
\item{flows}{A integer matrix of Origin-Destination flows.}

\item{distance}{a distance matrix between origins and destinations.}

\item{beta}{Exponent to use when calculating the cost function.}

\item{type}{The only type of cost function currently implemented is exponential, parameter value "exp".}
}
\value{
A list containing an integer matrix with predicted values.
}
\description{
This function is the C++ implementation of run_model, it will run a model
}
\examples{
a = 2
b = 3
a + b
}
