% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_rcpp.R
\name{simulation}
\alias{simulation}
\title{Running a whole simulation of a doubly constrained gravity model}
\usage{
simulation(
  flows_matrix,
  region_data,
  run_name,
  graph = NULL,
  from = NULL,
  to = NULL,
  graph_dist_matrix = NULL,
  beta_offset = 0,
  dist_threshold = 15000,
  norm = 2,
  time = FALSE,
  n_cores = 3,
  cost_fun = "exp"
)
}
\arguments{
\item{flows_matrix}{a integer matrix of flows}

\item{region_data}{a Sf data.frame containing at least a column which stores as well known text (wkt) the points of origin}

\item{run_name}{a name for the simulation}

\item{graph}{a cppRouting generated graph.}

\item{from}{the name of the culomn with origin points}

\item{to}{the name of the column with destination points, if same as from, set to NULL}

\item{graph_dist_matrix}{a network distance matrix in case it was already computed}

\item{beta_offset}{an offset from 0 from which to start looking for the best fit value.}

\item{dist_threshold}{a distance threshold to use.}

\item{norm}{kind of reference distance to use, 1 for manhattan, 2 for euclidean}

\item{time}{to convert distance to time, set to TRUE for time.}

\item{n_cores}{number of threads to run the model}

\item{cost_fun}{type of cost function}
}
\value{
creates a folder based on the run_name parameter to which images and files are written.
The file run_name_best_fit.rds contain the matrices with values from the model
, and the quality of fit values for the beta values.
}
\description{
this script takes a network, a flows data, and constructs everything from scratch to run a model based on some intput parameters.
}
