% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_rcpp.R
\name{run_model}
\alias{run_model}
\title{Running doubly constrained model}
\usage{
run_model(flows, distance, beta = 0.25, type = "exp")
}
\arguments{
\item{flows}{A integer matrix of Origin-Destination flows.}

\item{distance}{a distance matrix between origins and destinations, provide distance in km.}

\item{beta}{Exponent to use when calculating the cost function.}

\item{type}{The only type of cost function currently implemented is exponential, parameter value "exp".}
}
\value{
A list containing an integer matrix with predicted values.
}
\description{
This function is the C++ implementation of run_model, it will run a doubly constrained model
}
\examples{

data(flows_test)
data(distance_test)

model_test <- run_model(flows_test,distance_test)

}
